+++ aruw-mcb-project/src/aruwsrc/control/chassis/swerve_module.cpp
@@ -77,20 +77,21 @@ float SwerveModule::calculate(float x, float y, float r)
         newRotationSetpointRadians = newRawRotationSetpointRadians + rotationOffset;
 
         // normal angle wrapping
-        if (abs(newRotationSetpointRadians - preScaledRotationSetpoint) > M_PI)
+        if (abs(newRotationSetpointRadians - preScaledRotationSetpoint) > static_cast<float>(M_PI))
         {
-            rotationOffset -=
-                getSign(newRotationSetpointRadians - preScaledRotationSetpoint) * M_TWOPI;
+            rotationOffset -= getSign(newRotationSetpointRadians - preScaledRotationSetpoint) *
+                              static_cast<float>(M_TWOPI);
         }
         newRotationSetpointRadians = newRawRotationSetpointRadians + rotationOffset;
 
         // TODO: mechanical problem with the tension wheels in swerve module make this not work
         //       re-enable once fixed
         // reverse module if it's a smaller azimuth rotation to do so
-        if (abs(newRotationSetpointRadians - preScaledRotationSetpoint) > M_PI_2)
+        if (abs(newRotationSetpointRadians - preScaledRotationSetpoint) >
+            static_cast<float>(M_PI_2))
         {
-            rotationOffset -=
-                getSign(newRotationSetpointRadians - preScaledRotationSetpoint) * M_PI;
+            rotationOffset -= getSign(newRotationSetpointRadians - preScaledRotationSetpoint) *
+                              static_cast<float>(M_PI);
         }
         preScaledRotationSetpoint = newRawRotationSetpointRadians + rotationOffset;
 
@@ -100,7 +101,10 @@ float SwerveModule::calculate(float x, float y, float r)
         // if offset isn't an integer multiple of 2pi, it means module is currently reversed so
         // speed must be negative
         //  compareFloatClose may or may not be necessary
-        if (compareFloatClose(wrapAngle(rotationOffset, M_TWOPI), M_PI, 0.1))
+        if (compareFloatClose(
+                wrapAngle(rotationOffset, static_cast<float>(M_TWOPI)),
+                static_cast<float>(M_PI),
+                0.1))
             preScaledSpeedSetpoint *= -1;
     }
     return preScaledSpeedSetpoint;
